When rotating to portrait, also rotate the buffer.
This attempts to preserve existing behaviour by rotating in one
direction when going to portrait, and the other direction when going
to landscape. No rotation is done if the change in rotation is a
multiple of 180 degrees. This means that rotating from 0 to 180
degrees should behave the same as rotating from 0 to 90 and then
to 180. This is only important if you don't clear the display
when changing rotation.