Fix portrait modes.
Crop a portrait or landscape area based on the current rotation, and
then rotate it to the right orientation. This should work correctly
with non-standard sized buffers.

The initial buffer sizes are now calculated from max(width, height).
i.e. the initial buffer is square, and big enough to fit the physical
display in either orientation, but self.width and self.height always
refer to the physical display size.