Fix scrolling with transformations.
Introduces a new _grow_buffer function, which which replaces the
logic in set_pixel(). It is also now used in show(). Instead of a
straight copy, we start by growing the buffer so it is at least as
big as the (rotated) display. This means the buffer can start out
arbitrarily small, so there is no longer a problem with scrolling
through "empty space" in the pre-sized buffer.

The buffer now defaults to 1 pixel square.