Create wunderground-temp-display.py
Python script for the Raspberry Pi that pulls real-time weather data from Weather Underground, aka wunderground, and displays it on the scrollphathd. Includes the ability to specify location, temperature scale and whether you'd like the actual temperature or feels like temperature displayed. It also calculates an hourly running average and displays the temperature trend (up or down), and shows the current and gust wind speeds as a line graph on the bottom of the display, set for MPH if you're in a country using Fahrenheit and KPH if you're in a country using Celsius.  

I also added a fun "Knight Rider pulse" that shows that the code is running.

This code is placed in the public domain for all to enjoy.